fix(ui): align with master after icons consolidation and workspaces moves
- Update InfoOutlinedIcon imports to @/assets/icons per #19313 icons
  consolidation (was overwritten back to @/components/icons during
  rebase with -X theirs).
- Fix stale Header.tsx import path: SpacesSwitcher moved under
  ui/components/workspaces/ in #19311, so use ../../workspaces/SpacesSwitcher/SpaceSwitcher.
- Refresh eslint.config.js legacy allowlist entries for files moved
  by this PR (ConfirmationModal, ExportModal, ViewInfoModal,
  layout.tsx, navigatorComponents.tsx) and for workspaces paths now
  under components/workspaces/ rather than Lifecycle/Workspaces/.
