+++ ui/eslint.config.js
@@ -88,7 +88,7 @@ const legacyLiteralColorOffenders = [
   'components/Lifecycle/General/empty-state/curvedArrowIcon.tsx',
   'components/Lifecycle/General/empty-state/index.tsx',
   'components/Lifecycle/General/flip-card/index.tsx',
-  'components/Lifecycle/Workspaces/index.tsx',
+  'components/workspaces/index.tsx',
   'components/shared/LoadingState/Animations/AnimatedFilter.tsx',
   'components/shared/LoadingState/Animations/AnimatedLightMeshery.tsx',
   'components/shared/LoadingState/Animations/AnimatedMeshPattern.tsx',
@@ -183,7 +183,7 @@ const legacyInlineStyleOffenders = [
   'assets/icons/shapes/Octagon.tsx',
   'components/AppComponents.tsx',
   'components/BBChart.tsx',
-  'components/ConfirmationModal.tsx',
+  'components/shared/Modal/ConfirmationModal.tsx',
   'components/Dashboard/UnsavedChangesModal.tsx',
   'components/Dashboard/charts/ConnectionCharts.tsx',
   'components/Dashboard/charts/DashboardMeshModelGraph.tsx',
@@ -216,7 +216,7 @@ const legacyInlineStyleOffenders = [
   'components/designs/lifecycle/common.tsx',
   'components/designs/lifecycle/finalizeDeployment.tsx',
   'components/DuplicatesDataTable.tsx',
-  'components/ExportModal.tsx',
+  'components/shared/Modal/ExportModal.tsx',
   'components/FlipCard.tsx',
   'components/General/ConnectClustersBtn.tsx',
   'components/General/CreateDesignBtn.tsx',
@@ -229,10 +229,10 @@ const legacyInlineStyleOffenders = [
   'components/layout/Header/Header.tsx',
   'components/layout/Header/HeaderMenu.tsx',
   'components/Lifecycle/General/empty-state/index.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceActionList.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceDataTable.tsx',
-  'components/Lifecycle/Workspaces/WorkspaceGridView.tsx',
-  'components/Lifecycle/Workspaces/index.tsx',
+  'components/workspaces/WorkspaceActionList.tsx',
+  'components/workspaces/WorkspaceDataTable.tsx',
+  'components/workspaces/WorkspaceGridView.tsx',
+  'components/workspaces/index.tsx',
   'components/shared/LoadingState/Animations/AnimatedMeshSync.tsx',
   'components/shared/LoadingState/LoadingComponent.tsx',
   'components/shared/LoadingState/LoadingComponentServer.tsx',
@@ -316,7 +316,7 @@ const legacyInlineStyleOffenders = [
   'components/TroubleshootingComponent.tsx',
   'components/TypingFilter/index.tsx',
   'components/UserPreferences/index.tsx',
-  'components/ViewInfoModal.tsx',
+  'components/shared/Modal/ViewInfoModal.tsx',
   'components/ViewSwitch.tsx',
   'components/designs/configurator/MeshModel/LazyComponentForm.tsx',
   'components/designs/configurator/MeshModel/index.tsx',
@@ -337,9 +337,9 @@ const legacyInlineStyleOffenders = [
   'components/filters/Filters.tsx',
   'components/filters/FiltersCard.tsx',
   'components/filters/FiltersGrid.tsx',
-  'components/layout.tsx',
+  'components/layout/AppShell/layout.tsx',
   'components/multi-select-wrapper.tsx',
-  'components/navigatorComponents.tsx',
+  'components/layout/Navigator/navigatorComponents.tsx',
   'components/telemetry/grafana/GrafanaComponent.tsx',
   'components/telemetry/prometheus/PrometheusSelectionComponent.tsx',
   'pages/_app.tsx',
